Implement exclude categories from deletion

+++ app/src/main/java/eu/kanade/tachiyomi/ui/manga/MangaPresenter.kt
@@ -47,6 +47,7 @@ import exh.metadata.metadata.base.getFlatMetadataForManga
 import exh.source.EnhancedHttpSource.Companion.getMainSource
 import exh.util.asObservable
 import exh.util.await
+import exh.util.shouldDeleteChapters
 import exh.util.trimOrNull
 import kotlinx.coroutines.NonCancellable
 import kotlinx.coroutines.withContext
@@ -807,7 +808,7 @@ class MangaPresenter(
         launchIO {
             db.updateChaptersProgress(chapters).executeAsBlocking()
 
-            if (preferences.removeAfterMarkedAsRead()) {
+            if (preferences.removeAfterMarkedAsRead() /* SY --> */ && manga.shouldDeleteChapters(db, preferences) /* SY <-- */) {
                 deleteChapters(chapters)
             }
         }
